net/ena: fix L4 cksum flags condition for TX
During an if-condition evaluation, a 2-bit flag evaluates to 'true' for
'0x1', '0x2' and '0x3'. Thus, from this perspective these flags are
indistinguishable. To make them distinct, respective bits must be
extracted with a mask and then checked for strict equality.

Specifically here, even if `PKT_TX_UDP_CKSUM` (value '0x3') was set, the
expression `mbuf->ol_flags & PKT_TX_TCP` (the second flag of value
'0x1') is evaluated first and the result is 'true'. In consequence, for
UDP packets the execution flow enters an incorrect branch.

Fixes: 56b8b9b7e5d2 (net/ena: convert to new Tx offloads API)
Cc: stable@dpdk.org

Signed-off-by: Maciej Bielski <mba@semihalf.com>
Reported-by: Eduard Serra <eduser25@gmail.com>
Acked-by: Michal Krawczyk <mk@semihalf.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>