doc: announce API change in mbuf
In order to prepare for a long term stable API, the mbuf library
has to change: allowing more features as dynamic fields,
and fixing the lack of rte_ prefix in the namespace.
The namespace fix should not break the compatibility by keeping
some aliases during few releases to give time for migration.

Signed-off-by: Thomas Monjalon <thomas@monjalon.net>
Signed-off-by: 0-day Robot <robot@bytheb.org>