net/mlx5: fix order of lines in loop
Function flow_dv_zero_encap_udp_csum() uses a while loop to iterate
over vlan items in flow rule.
Pointer next_hdr is incremented to the next item before it is used,
so the first item is skipped.

This patch moves the incrementing of next_hdr to the correct place.

Fixes: bf1d7d9a033a ("net/mlx5: zero out UDP checksum in encapsulation")
Cc: elibr@mellanox.com

Signed-off-by: Dekel Peled <dekelp@mellanox.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>