doc: announce new mbuf field for LRO
The API breakage is because the ``tso_segsz`` field was documented for
LRO.

The ``tso_segsz`` field in mbuf indicates the size of each segment in
the LRO packet in Rx path and should be provided by the LRO packet
port.

While the generic LRO packet may aggregate different segments sizes in
one packet, it is impossible to expose this information for each segment
by one field and it doesn't make sense to expose all the segments sizes
in the mbuf.

A new field may be added as union with the above field to expose the
number of segments aggregated in the LRO packet.

Signed-off-by: Matan Azrad <matan@mellanox.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>