cryptodev: fix invalid dev_id after a pmd close
Each cryptodev are indexed with its dev_id in the global
rte_crypto_devices variable. nb_devs is incremented / decremented each
time a cryptodev is created / deleted. The goal of nb_devs is to prevent
the user to get an invalid dev_id.

Let's imagine DPDK has configured N cryptodevs. If the cryptodev=1 is
removed at runtime, the latest cryptodev N cannot be accessible, because
nb_devs=N-1.

In order to prevent this kind of behavior, let's remove the check with
nb_devs and iterate in all the rte_crypto_devices elements: if data is
not NULL, that means a valid cryptodev is available.

Fixes: d11b0f30df88 ("cryptodev: introduce API and framework for crypto devices")
Cc: stable@dpdk.org

Signed-off-by: Julien Meunier <julien.meunier@nokia.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>