drivers/crypto/openssl: use a local copy for the session contexts
Session contexts are used for temporary storage when processing a
packet.
If packets for the same session are to be processed simultaneously on
multiple cores, separate contexts must be used.

Note: with openssl 1.1.1 EVP_CIPHER_CTX can no longer be defined as a
variable on the stack: it must be allocated. This in turn reduces the
performance.

Fixes: d61f70b4c918 ('crypto/libcrypto: add driver for OpenSSL library')
Cc: stable@dpdk.org
Cc: declan.doherty@intel.com

Signed-off-by: Thierry Herbelot <thierry.herbelot@6wind.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>