vhost: translate incoming log address to gpa
When IOMMU is enabled the incoming log address is in IOVA space. In that
case, look in IOTLB table and translate the resulting HVA to GPA.

If IOMMU is not enabled, the incoming log address is already a GPA so no
transformation is needed.

This change makes page logging work when IOVA_VA is selected in the guest.
Further information: https://bugs.dpdk.org/show_bug.cgi?id=337

Cc: Maxime Coquelin <maxime.coquelin@redhat.com>
Reported-by: Pei Zhang <pezhang@redhat.com>

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>