eal:ppc: fix incorrect ifdef for ppc_64
An ifdef present in eal_memory.c references "RTE_ARCH_PPC64" when
it should actually use "RTE_ARCH_PPC_64".  Simple testing revealed
that both the PPC_64 and non-PPC_64 versions of the code involved
work, but the PPC_64 version of the code is retained to be
consistent with other instances in the same file where mmapped
memory is accessed in reverse order on Power platforms.

Fixes: 66cc45e ("mem: replace memseg with memseg lists")
Cc: Anatoly Burakov <anatoly.burakov@intel.com>

Signed-off-by: David Christensen <drc@linux.vnet.ibm.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>