net/mlx5: fix the alloc size of rqt attribute
The receive queues list size is based on the size of uint32_t, so
when allocating the memory, the correct value should be used. Or
else there is risk to corrupt the memory, depending on the queues
number, because there is some pad area for alignment. If the queue
number is not large enough, the issue couldn't be observed.

Based on Bing fix.

Fixes: dc9ceff73c99 ("net/mlx5: create advanced RxQ via DevX")
Cc: dekelp@mellanox.com

Signed-off-by: Ori Kam <orika@mellanox.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>