examples/quota-watermark: remove example from DPDK
Original DPDK rings code had explicit support for a single watermark
per-ring, but more recent releases of DPDK had a more general mechanism
where each enqueue or dequeue call could return the remaining
elements/free-slots in the ring.  Therefore, this example is not as
relevant as before and can be removed.

Signed-off-by: Ciara Power <ciara.power@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>