net/af_packet: improve Tx statistics accuracy
When sendto call fails and ENOBUFS error is being set some of the
packets are actually successfully transmitted. There is no available
count of those packets, so in order to make the statistics more
accurate, all the previously enqueued packets will be considered
successful, even though this is not entirely correct.

Before:
   testpmd Tx statistics:
      TX-packets: 7529062   TX-errors: 3702150  TX-bytes:  451743720
   pktgen Rx statistics:
      Total Rx Pkts: 10700700

After:
   testpmd TX statistics:
      TX-packets: 11510625  TX-errors: 0        TX-bytes:  690637500
   pktgen Rx statistics:
      Total Rx Pkts: 10974307

Fixes: 74b7fc0a0ff1 ("net/af_packet: fix packet bytes counting")
Cc: ciwillia@brocade.com
Cc: stable@dpdk.org

Signed-off-by: Flavia Musatescu <flavia.musatescu@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>