ebpf issues with secondary process.
Trying to use the existing BPF code in BPF for packet filter is possible.
But one road block is that capture wants to allow a secondary process to
insert a packet filter. The current code in bpf_load() won't work for that.
The BPF program needs to be in shared area to allow secondary process
to work.

Why does the code use mmap() as an allocator instead of rte_malloc?

Something like the diff below (untested). JIT would have the same problem
but you would have issues with the mprotect stuff with huge pages.

I also noticed that bpf code is not using rte_memcpy and it has unnecessary
casts to void *.

Signed-off-by: 0-day Robot <robot@bytheb.org>