doc: disable doxygen's HAVE_DOT tag
Ubuntu ships with a patched version of doxygen that enables
HAVE_DOT. Enabling this option causes the warning:
"""
warning: Included by graph for 'rte_common.h' not generated,
too many nodes. Consider increasing DOT_GRAPH_MAX_NODES
"""
This reproduces with doxygen version 1.8.13 and
dot - graphviz version 2.40.1 on Ubuntu 18.04.

This will force doxygen not to assume that dot (part of Graphviz)
is installed, which is the default, and will result in dot not being used
for visualization. The other option is to increase DOT_GRAPH_MAX_NODES
to a value larger than 55.

See:
- http://changelogs.ubuntu.com/changelogs/pool/main/d/doxygen/doxygen_1.8.13-10/changelog
- http://www.doxygen.nl/manual/config.html#cfg_have_dot

Signed-off-by: Ali Alnubani <alialnu@mellanox.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>