cmdline: prefix cmdline numeric enum
The values in an enum are really global names and not specific
to the enum in question. This can lead to namespace conflicts
in applications if a common value is visible.

The DPDK cmdline API has an enum for numeric type values with
names line UINT32 which could be used or defined in a user
application. Change these to be prefixed with the enum name
like other places in DPDK.

Lots of lines changed with no change in actual code
generated.

Signed-off-by: Stephen Hemminger <stephen@networkplumber.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>