net/ice: use dynamic mbuf API to handle protocol extraction
The original design is to occupy rte_mbuf::udata64 to save the metadata
which combines the network protocol fields and related type, then use
private API to decode this metadata. It is no so generic.

Use the new dynamic mbuf field and flags to handle protocol extraction.
Now the new metadata is very clean, just network protocol fields in the
specified offset, and its type will be indicated by related ol_flags.

Signed-off-by: Haiyue Wang <haiyue.wang@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>