net/ice: optimize protocol extraction by dynamic mbuf API
The original design is to use rte_mbuf::udata64 to save the metadata of
protocol extraction which has network protocol data fields and type, an
private API is used to decode this metadata. It is not so generic.

Use the new dynamic mbuf field and flags to handle protocol extraction.
Then the metadata is very clean, only network protocol fields extracted
by defined offset, and its type will be indicated by related ol_flags.

Signed-off-by: Haiyue Wang <haiyue.wang@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>