net/ice/base: fix wrong symmetric hash configure
When a new hash profie is created, we need to reset all related
GLQF_HSYMM registers, otherwise unexpected hash behaviour may happens
on packet that hit that profile.

The patch fix the issue that we only do reset when symmetric hash
is required, but actually for non symmmetric hash we also need this,
since GLQF_HSYMM might be polluted by previous configuration.

Fixes: ddae0440353f ("net/ice/base: enable symmetric hash for RSS")

Signed-off-by: Qi Zhang <qi.z.zhang@intel.com>
Tested-by: Simei Su <simei.su@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>