ethdev: fix last item detection on RSS flow expand
There is a rte_flow API which expands a RSS flow pattern to multipile
patterns according to the RSS hash types in the RSS action
configuration.

Aa part of the expansion, detection of the last item of the flow uses
the "next proto" field of the last configured item in the pattern list.
Wrongly, the mask of this field was not considered in order to validate
the field.

Ignore "next proto" fields when their corresponded masks invalidate them.

Fixes: ec84aa45f17b ("ethdev: fix expand RSS flows")
Cc: jackmin@mellanox.com

Signed-off-by: Matan Azrad <matan@mellanox.com>
Acked-by: Xiaoyu Min <jackmin@mellanox.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>